Kind of an early POC on new --platform
libretro-super.sh works, but you currently need to set platform and ARCH
in your environment before running it as you do with libretro-build.sh
now (see libretro-build-ios.sh wrapper for what I mean).  The major work
left here is the --platform argument and a bunch of little things that
will mostly fall into place when that's done.  Also need to settle on a
license for this.

My habit for scripts is public domain, but it could be argued that
libretro-super is no longer "a script".  Still, I want people to use
this if it's useful to them, so I'd prefer a totally unencumbered
license in the vein of MIT/BSD/CC-BY.  I've tentatively selected CC-BY
as a RFC on it vs. MIT, mostly because it tries to explain things in
plain language and unlike MIT/BSD is reviewed in the context of
(d)evolving international laws to ensure that it still does what is
intended.