libretro-buildbot-recipe.sh: Always use the branch in the recipe file.
If the checked out branch does not match the branch name in the recipe
file the script will now make a shallow fetch of the correct branch,
checkout the new branch and then remove the old one. If this is done it
will also build a core with the new branch.