libretro-buildbot-recipe.sh: Fix builds with force pushes.
The script will no longer git pull, but will fetch the any changes
from the remote branch and then reset the repo to that branch.

If there are changes the HEAD commit will change and it will rebuild.

This will also work if the old history was changed instead of the new
history.