libretro-buildbot-recipe.sh: Fix constant rebuilds for some cores.
A few cores built with the LEIRADEL command had an issue where a
hack used to build certain cores broke and caused them to be built
every time the buildbot encountered the recipe.

To fix this the build_libretro_leiradel_makefile function was removed
and incorporated into the build_libretro_generic_makefile function.
This allowed the script to easily use new logic to build multiple cores
with one recipe which entirely avoids the need for the broken hack.