recipes: Fix constant rebuilds for thepowderoy.
Currently thepowdertoy shares the build directory for three
different android builds, aarch64, armv7 and x86. However the
recipes for these builds differ and thepowdertoy will constantly
rebuild when it finds the recipe file differs with the previous
build.

This can be avoid by appending the arch to the build directory and
this solution is already used by ppsspp in the same recipes.