remove "experimental" from mame2003-plus name, etc
When I wrote the original .info file for mame2003-plus, the builds were much less stable and we knew we were about to change the input system.

Now it is about as stable as any other core that gets updates at all. I'd like to remove Experimental and standardize a few other lines.