Fix corename for mame2003
== DETAILS

A recent change in MAME 2003 changed the internal corename to be
simply "mame2003". This ended up breaking netplay, because netplay
uses that internal name to map to the core filename; and with the
name no longer matching, the core lookup fails.

I've pitched a PR to revert that change to the internal core name,
but it was rejected. So this is the next safest way to fix the problem.
And really, it needs to change because the version string in the
internal name was redundant anyway.