Add a check for Dolphin system folder.
Check internally for `system/dolphin-emu/Sys/codehandler.bin` (a file that should be here if the folder was set up properly), for the user it will just tell if "Sys" folder is present or not.

It will be useful for troubleshooting, because we'll know directly if the folder is there AND properly set up/not empty.