Update freeJ2ME's core info file
Retroarch has a database for "Mobile - J2ME", which is precisely
what FreeJ2ME supports. It also sports very little libretro
functionality so this should be noted in the info file, and it's
"BIOS/Firmware" file does not have a fixed md5 due to that file
actually being the emulator itself.